weird things happen when going between desktop and mobile
