+++ src/components/layouts/MobileLayout.js
@@ -1,0 +1,12 @@
+
+
+import React from 'react';
+import { Text } from '@chakra-ui/react';
+
+export default function MobileLayout(companyCategories, handleSelectedCompany, outerBounds, setSelectedCompany, setMap, companies, selectedCompany) {
+    return (
+        <>
+            <Text>Mobile Layout</Text>
+        </>
+    )
+}
